delete popper.js and add floating-ui vendor js
